Move `circuit-disband` tests into separate module
This change moves the admin service integration tests that validated the
process of creating, then disbanding a circuit into a separate module.
This makes the tests easier to read, as they are separated by concern in
each module.
